Macro pallet_ovm::try_or_exec_error

macro_rules! try_or_exec_error {
    ($e:expr, $buffer:expr) => { ... };
}

Evaluate an expression of type Result<_, &'static str> and either resolve to the value if Ok or wrap the error string into an ExecutionError with the provided buffer and return from the enclosing function. This macro is used instead of .map_err(..)? in order to avoid taking ownership of buffer unless there is an error.