Module pallet_ovm::predicate

Structs

CallContext

ExecError
An error indicating some failure to execute a contract call or instantiation. This can include VM-specific errors during execution (eg. division by 0, OOB access, failure to satisfy some precondition of a system call, etc.) or errors with the orchestration (eg. out-of-gas errors, a non-existent destination contract, etc.).

ExecutionContext

OvmExecutable
Ovm executable loaded by OvmLoader and executed by OptimisticOvm.

PredicateLoader
Loader which fetches OvmExecutable from the code cache.

PredicateOvm
Implementation of Vm that takes PredicateOvm and executes it.

PrefabOvmModule
A prepared wasm module ready for execution.

Enums

PredicateError
Reason why a predicate call failed

Functions

save_code
Put code in the storage. The hash of code is used as a key and is returned as a result of this function.

Type Definitions

ExecResult