Function pallet_ovm::predicate::save_code

pub fn save_code<T: Trait>(
    original_code: Vec<u8>, 
    schedule: &Schedule
) -> Result<<T as Trait>::Hash, &'static str>

Put code in the storage. The hash of code is used as a key and is returned as a result of this function.

This function instruments the given code and caches it in the storage.