Struct pallet_ovm::Module

pub struct Module<T: Trait>(_);

Implementations

impl<T: Trait + 'static> Module<T>

pub fn current_schedule() -> Schedule

Current cost schedule for contracts.

pub fn predicate_codes<K: EncodeLike<<T as Trait>::Hash>>(
key: K
) -> Option<Vec<u8>>

A mapping from an original code hash to the original code, untouched by instrumentation.

pub fn predicate_cache<K: EncodeLike<<T as Trait>::Hash>>(
key: K
) -> Option<PrefabOvmModule>

A mapping between an original code hash and instrumented ovm(predicate) code, ready for execution.

pub fn predicates<K: EncodeLike<T::AccountId>>(
key: K
) -> Option<PredicateContract<<T as Trait>::Hash>>

Mapping the predicate address to Predicate. Predicate is handled similar to contracts.

pub fn games<K: EncodeLike<T::Hash>>(
key: K
) -> Option<ChallengeGame<<T as Trait>::Hash, <T as Trait>::BlockNumber>>

Mapping the game id to Challenge Game.

impl<T: Trait> Module<T>

Can also be called using Call.

pub fn put_code(origin: T::Origin, predicate: Vec<u8>) -> DispatchResult

Stores the given binary Wasm code into the chain's storage and returns its codehash. You can instantiate contracts only with stored code.

NOTE: Calling this function will bypass origin filters.

pub fn instantiate(
origin: T::Origin,
predicate_hash: <T as Trait>::Hash,
inputs: Vec<u8>
) -> DispatchResult

Deploy predicate and made predicate address as AccountId. TODO: weight

NOTE: Calling this function will bypass origin filters.

pub fn claim(origin: T::Origin, claim: PropertyOf<T>) -> DispatchResult

Claims property and create new game. Id of game is hash of claimed property TODO: weight

NOTE: Calling this function will bypass origin filters.

pub fn challenge(
origin: T::Origin,
property: PropertyOf<T>,
challenge_property: PropertyOf<T>
) -> DispatchResult

Challenge to an existing game instance by a property.

challenge will be added to challenges field of challenged game instance. if property does not exist, revert. if challenge with same property was made before, revert.

TODO: weight

NOTE: Calling this function will bypass origin filters.

pub fn remove_challenge(
origin: T::Origin,
property: PropertyOf<T>,
challenge_property: PropertyOf<T>
) -> DispatchResult

remove challenge set challenging game decision to false and remove it from challenges field of challenged game if property does not exist, revert. if challenge property does not exist, revert.

TODO: weight

NOTE: Calling this function will bypass origin filters.

pub fn set_game_result(
origin: T::Origin,
property: PropertyOf<T>,
result: bool
) -> DispatchResult

set game result to given result value. only called from dispute contract

TODO: weight

NOTE: Calling this function will bypass origin filters.

pub fn settle_game(origin: T::Origin, property: PropertyOf<T>) -> DispatchResult

settle game settle started game whose dispute period has passed. if no challenge for the property exists, decide to true. if any of its challenges decided to true, decide game to false. if undecided challenge remains, revert.

TODO: weight

NOTE: Calling this function will bypass origin filters.

impl<T: Trait> Module<T>

pub fn bare_call(
origin: T::AccountId,
dest: T::AccountId,
input_data: Vec<u8>
) -> ExecResult<T>

Perform a call to a specified contract.

This function is similar to Self::call, but doesn't perform any address lookups and better suitable for calling directly from Rust.

pub fn is_decided(property: &PropertyOf<T>) -> Decision

Get of true/false the decision of property.

pub fn is_decided_by_id(id: T::Hash) -> Decision

Get of true/false the decision of game id.

pub fn get_game(
claim_id: &T::Hash
) -> Option<ChallengeGame<<T as Trait>::Hash, <T as Trait>::BlockNumber>>

Get of the instatiated challenge game from claim_id.

pub fn get_property_id(property: &PropertyOf<T>) -> T::Hash

Get of the property id from the propaty itself.

pub fn is_challenge_of(
property: &PropertyOf<T>,
challenge_property: &PropertyOf<T>
) -> bool

pub fn started(id: &T::Hash) -> bool

check if game of given id is already started.

pub fn block_number() -> <T as Trait>::BlockNumber

pub fn is_decidable(property_id: &T::Hash) -> bool