Trait pallet_ovm::Trait

pub trait Trait: Trait {
    type MaxDepth: Get<u32>;
    type DisputePeriod: Get<Self::BlockNumber>;
    type DeterminePredicateAddress: PredicateAddressFor<Self::Hash, Self::AccountId>;
    type HashingL2: Hash<Output = Self::Hash>;
    type ExternalCall: Ext<Self> + NewCallContext<Self>;
    type AtomicPredicateIdConfig: Get<AtomicPredicateIdConfig<Self::AccountId, Self::Hash>>;
    type Event: From<Event<Self>> + Into<Self::Event>;
}

Associated Types

type MaxDepth: Get<u32>

The maximum nesting level of a call/instantiate stack.

type DisputePeriod: Get<Self::BlockNumber>

During the dispute period defined here, the user can challenge. If nothing is found, the state is determined after the dispute period.

type DeterminePredicateAddress: PredicateAddressFor<Self::Hash, Self::AccountId>

A function type to get the contract address given the instantiator.

type HashingL2: Hash<Output = Self::Hash>

The hashing system (algorithm) being used in the runtime (e.g. Keccak256).

type ExternalCall: Ext<Self> + NewCallContext<Self>

ExternalCall context.

type AtomicPredicateIdConfig: Get<AtomicPredicateIdConfig<Self::AccountId, Self::Hash>>

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

Loading content...