Module pallet_ovm::traits

Traits

Ext
An interface that provides access to the external environment in which the predicate-contract is executed similar to a smart-contract.

Loader
Loader is a companion of the Vm trait. It loads an appropriate abstract executable to be executed by an accompanying Vm implementation.

PredicateAddressFor
A function that generates an AccountId for a predicate upon instantiation.

Vm
A trait that represent an optimistic virtual machine.