Trait pallet_ovm::traits::Ext

pub trait Ext {
    type T: Trait;
    fn call(
        &self, 
        to: &<Self::T as Trait>::AccountId, 
        input_data: Vec<u8>
    ) -> ExecResult;

    fn caller(&self) -> &<Self::T as Trait>::AccountId;

    fn address(&self) -> &<Self::T as Trait>::AccountId;
}

An interface that provides access to the external environment in which the predicate-contract is executed similar to a smart-contract.

This interface is specialized to an account of the executing code, so all operations are implicitly performed on that account.

Predicate call the only below functions.

call: call to other predicate.

caller: get of the caller of this predicate.

address: the predicate's address.

is_stored: check the storage of other modules or contracts.

Associated Types

type T: Trait

Loading content...
Required methods

fn call(
&self,
to: &<Self::T as Trait>::AccountId,
input_data: Vec<u8>
) -> ExecResult

Call (possibly other predicate) into the specified account.

fn caller(&self) -> &<Self::T as Trait>::AccountId

Returns a reference to the account id of the caller.

fn address(&self) -> &<Self::T as Trait>::AccountId

Returns a reference to the account id of the current contract.

Loading content...