Trait pallet_ovm::traits::Loader

pub trait Loader<T: Trait> {
    type Executable;
    fn load_main(
        &self, 
        code_hash: &<T as Trait>::Hash
    ) -> Result<Self::Executable, &'static str>;
}

Loader is a companion of the Vm trait. It loads an appropriate abstract executable to be executed by an accompanying Vm implementation.

Associated Types

type Executable

Loading content...
Required methods

fn load_main(
&self,
code_hash: &<T as Trait>::Hash
) -> Result<Self::Executable, &'static str>

Load the main portion of the code specified by the code_hash. This executable is called for each call to a contract.

Loading content...