Trait pallet_ovm::traits::Vm

pub trait Vm<T: Trait> {
    type Executable;
    fn execute(
        &self, 
        exec: Self::Executable, 
        ext: T::ExternalCall, 
        input_data: Vec<u8>
    ) -> ExecResult<T>;
}

A trait that represent an optimistic virtual machine.

You can view an optimistic virtual machine as something that takes code, an input data buffer, queries it and/or performs actions on the given Ext and optionally returns an output data buffer. The type of code depends on the particular virtual machine.

Execution of code can end by either implicit termination (that is, reached the end of executable), explicit termination via returning a buffer or termination due to a trap.

Associated Types

type Executable

Loading content...
Required methods

fn execute(
&self,
exec: Self::Executable,
ext: T::ExternalCall,
input_data: Vec<u8>
) -> ExecResult<T>

Loading content...