Trait pallet_ovm::traits::Ext

pub trait Ext<T: Trait> {
    fn call(
        &self, 
        to: &<T as Trait>::AccountId, 
        input_data: Vec<u8>
    ) -> ExecResult<T>;

    fn caller(&self) -> &<T as Trait>::AccountId;

    fn address(&self) -> &<T as Trait>::AccountId;

    fn is_stored(
        &self, 
        _address: &<T as Trait>::AccountId, 
        _key: &[u8], 
        _value: &[u8]
    ) -> bool;

    fn verify_inclusion_with_root(
        &self, 
        leaf: T::Hash, 
        token_address: <T as Trait>::AccountId, 
        range: &[u8], 
        inclusion_proof: &[u8], 
        root: &[u8]
    ) -> bool;

    fn is_decided(&self, property: &PropertyOf<T>) -> bool;

    fn is_decided_by_id(&self, id: T::Hash) -> bool;

    fn set_predicate_decision(
        &self, 
        game_id: T::Hash, 
        decision: bool
    ) -> Result<bool, ExecError<T::AccountId>>;
}

An interface that provides access to the external environment in which the predicate-contract is executed similar to a smart-contract.

This interface is specialized to an account of the executing code, so all operations are implicitly performed on that account.

Predicate call the only below functions.

call: call to other predicate.

caller: get of the caller of this predicate.

address: the predicate's address.

is_stored: check the storage of other modules or contracts.

Required methods

fn call(
&self,
to: &<T as Trait>::AccountId,
input_data: Vec<u8>
) -> ExecResult<T>

Call (possibly other predicate) into the specified account.

fn caller(&self) -> &<T as Trait>::AccountId

Returns a reference to the account id of the caller.

fn address(&self) -> &<T as Trait>::AccountId

Returns a reference to the account id of the current contract.

fn is_stored(
&self,
_address: &<T as Trait>::AccountId,
_key: &[u8],
_value: &[u8]
) -> bool

fn verify_inclusion_with_root(
&self,
leaf: T::Hash,
token_address: <T as Trait>::AccountId,
range: &[u8],
inclusion_proof: &[u8],
root: &[u8]
) -> bool

Needs Plasma.

fn is_decided(&self, property: &PropertyOf<T>) -> bool

fn is_decided_by_id(&self, id: T::Hash) -> bool

fn set_predicate_decision(
&self,
game_id: T::Hash,
decision: bool
) -> Result<bool, ExecError<T::AccountId>>

Loading content...